Pull apart the AssetsBuilder
